fix: logica de vuelos, aeropuertos y aerolineas

+++ src/main/java/com/airline/airline/dto/PasajeroDTO.java
@@ -4,6 +4,11 @@ import com.airline.airline.entities.Reserva;
 
 import java.util.Date;
 
-public record PasajeroDTO(Long id, String nombre, String apellido, String tipoDocumento, Integer numeroDocumento,
-                          Date fechaNacimiento, Long reserva) {
+public record PasajeroDTO(Long id,
+                          String nombre,
+                          String apellido,
+                          String tipoDocumento,
+                          Long numeroDocumento,
+                          Date fechaNacimiento,
+                          Long reserva) {
 }
